Update functions-app-settings.md
Update for app settings: WEBSITE\_CONTENTAZUREFILECONNECTIONSTRING WEBSITE\_CONTENTSHARE
There is a confusion for customers regarding Premium, as it refers to Elastic Premium and customers may think it is regarding Premium App Service plans. 
I believe description on this paragraph should be as follows:
For Consumption & Elastic Premium plans only. Connection string for storage account where the function app code and configuration are stored.